gcp_deployment/requirements.txt
# Core Agent-Based Modeling
mesa>=2.2.0
# mesa-frames>=0.1.0a0  # Install from GitHub instead

# LLM and ML
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
accelerate>=0.20.0
datasets>=2.12.0

# Atropos (install from git)
# atropos-rl  # Will install from GitHub

# 3D Asset Generation
point-e  # Will install from OpenAI GitHub
trimesh>=3.15.0
numpy>=1.21.0
Pillow>=9.0.0

# Performance & Monitoring
psutil>=5.9.0
memory-profiler>=0.60.0
line-profiler>=4.0.0
tqdm>=4.64.0

# Data Management & Storage
pandas>=1.5.0
sqlalchemy>=2.0.0
# sqlite3  # Built-in Python module

# Async & Networking
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Utilities
pydantic>=2.0.0
rich>=13.0.0
typer>=0.9.0
PyYAML>=6.0.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Dependencies for your research pipeline classes
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.17.0
networkx>=3.2.1
sympy>=1.12
wandb>=0.16.1
openai>=1.3.7
anthropic>=0.8.1
google-generativeai>=0.3.2
scikit-learn>=1.3.0
scipy>=1.11.4
python-dateutil>=2.8.2
pytz>=2023.3

# Add any other direct pip dependencies here
# For example, if your llm_services.py used other specific libraries
